IntelEdgeSolution/IntelVision/modules/VisionSampleModule/requirements.txt
azure-iothub-device-client
setuptools 
numpy 

opencv-python

pillow